findByEmail has been moved to User class.
