+++ src/Blog/Model/User.php
@@ -10,6 +10,7 @@ namespace Blog\Model;
 
 use Framework\Model\ActiveRecord;
 use Framework\Security\Model\UserInterface;
+use Framework\DI\Service;
 
 class User extends ActiveRecord implements UserInterface
 {
@@ -27,4 +28,22 @@ class User extends ActiveRecord implements UserInterface
     {
         return $this->role;
     }
+
+    /**
+     * Find records in database table by email attribute
+     *
+     * @param $email
+     * @return mixed
+     */
+    public static function findByEmail($email)
+    {
+        $db = Service::get('db');
+        $table = static::getTable();
+        $sql = "SELECT * FROM " . $table . " WHERE email = ?";
+        $stmt = $db->prepare($sql);
+        $stmt->execute(array((string)$email));
+        $result = $stmt->fetchAll(\PDO::FETCH_CLASS, get_called_class());
+        return $result[0];
+    }
+
 }
